#1 Убрал всё лишнее и оформил код
