--- license_plate_recognition_results.csv
@@ -1,31 +1,0 @@
-4,140.4807891845703,473.1668395996094,292.8459167480469,510.5116882324219,E074YH76
-7,158.17723083496094,481.0433044433594,316.3096923828125,518.0162353515625,E074XY76
-19,169.01637268066406,491.6084289550781,334.1131286621094,530.2810668945312,E074YH76
-27,96.72474670410156,302.936279296875,225.5789337158203,329.90325927734375,H878EP15
-29,96.58671569824219,301.7009582519531,226.49093627929688,329.8056335449219,H878EP05
-44,97.23820495605469,236.1718292236328,197.30551147460938,257.9277648925781,K048BY790
-45,98.19483947753906,236.77978515625,197.2119903564453,258.34039306640625,K048BY790
-46,98.19387817382812,236.77951049804688,197.21389770507812,258.34075927734375,K048BY790
-47,97.88484954833984,236.8786163330078,197.48345947265625,258.5937194824219,K048BY790
-48,97.8832778930664,236.87899780273438,197.4836883544922,258.5953063964844,K048BY790
-49,93.40643310546875,247.92626953125,194.4711456298828,270.09765625,K048BY790
-50,100.53755187988281,245.59608459472656,201.27911376953125,268.25885009765625,K048BY790
-51,101.99359130859375,245.86819458007812,203.41290283203125,267.8917541503906,X048BY790
-56,105.7992172241211,244.65890502929688,211.69247436523438,267.6390380859375,K048BY790
-70,103.982666015625,264.5262451171875,220.36114501953125,290.35577392578125,K048BY790
-71,111.14949798583984,263.92913818359375,228.53256225585938,290.1294860839844,K048BY790
-72,112.884521484375,266.5674743652344,231.8282928466797,291.9700622558594,K048BY790
-73,114.9057388305664,267.62200927734375,232.75198364257812,292.0274353027344,K048BY790
-74,114.98660278320312,264.1390686035156,231.62356567382812,291.07373046875,K048BY790
-75,115.11038970947266,263.9680480957031,231.52125549316406,290.5610046386719,K048BY790
-76,113.95097351074219,264.0339050292969,231.37998962402344,290.4419860839844,K048BY790
-77,113.18761444091797,265.5746154785156,231.58409118652344,292.5786437988281,K048BY790
-78,113.16943359375,265.5720520019531,231.5857391357422,292.5842590332031,K048BY790
-81,117.01683044433594,276.160400390625,237.07012939453125,302.6688537597656,K048BY780
-83,118.59093475341797,274.1480407714844,237.99072265625,300.32733154296875,K048BY790
-85,118.57009887695312,274.1808166503906,238.0001220703125,300.33428955078125,K048BY790
-87,117.04069519042969,275.21783447265625,237.0997314453125,301.844482421875,K048BY780
-90,118.0522232055664,290.518310546875,239.62802124023438,317.53350830078125,X048BY790
-97,109.13066101074219,307.6358642578125,235.07586669921875,335.6237487792969,X048BY70
-98,119.8558578491211,304.08447265625,244.85154724121094,330.715087890625,X048EX790
-99,124.5237045288086,304.0000915527344,249.12579345703125,330.74066162109375,X048BY70
